Envio dos cadastros para a nuvem, controle de ativação e sincronia implementados com sucesso.

+++ app/src/main/java/com/example/rcksuporte05/rcksistemas/fragment/CadastroCliente8.java
@@ -1,5 +1,7 @@
 package com.example.rcksuporte05.rcksistemas.fragment;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
@@ -24,13 +26,14 @@ public class CadastroCliente8 extends Fragment {
     EditText edtobsFat;
     @BindView(R.id.edtObsFinancas)
     EditText edtObsFinancas;
+    private DBHelper db;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.activity_cadastro_cliente8, container, false);
         ButterKnife.bind(this, view);
 
-        if (getActivity().getIntent().getIntExtra("vizualizacao", 0) >= 1) {
+        if (getActivity().getIntent().getIntExtra("vizualizacao", 0) >= 1 || ClienteHelper.getCliente().getId_cadastro_servidor() > 0) {
 
             edtobsFat.setFocusable(false);
             edtObsFinancas.setFocusable(false);
@@ -41,20 +44,44 @@ public class CadastroCliente8 extends Fragment {
                 @Override
                 public void onClick(View v) {
                     try {
-                        DBHelper db = new DBHelper(getActivity());
+                        db = new DBHelper(getActivity());
                         inserirDadosDaFrame();
 
                         if (ClienteHelper.salvarCliente()) {
-                            if (ClienteHelper.getCliente().getId_cadastro() <= 0)
-                                db.inserirTBL_CADASTRO(ClienteHelper.getCliente());
-                            else
-                                db.atualizarTBL_CADASTRO(ClienteHelper.getCliente());
+                            if (verificaCpfCnpj(ClienteHelper.getCliente().getCpf_cnpj())) {
+                                if (ClienteHelper.getCliente().getId_cadastro() <= 0)
+                                    db.inserirTBL_CADASTRO(ClienteHelper.getCliente());
+                                else
+                                    db.atualizarTBL_CADASTRO(ClienteHelper.getCliente());
 
-                            if (getActivity().getIntent().getIntExtra("prospect", 0) > 0)
-                                db.alterar("DELETE FROM TBL_PROSPECT WHERE ID_PROSPECT = " + getActivity().getIntent().getIntExtra("prospect", 0) + ";");
+                                if (getActivity().getIntent().getIntExtra("prospect", 0) > 0)
+                                    db.alterar("DELETE FROM TBL_PROSPECT WHERE ID_PROSPECT = " + getActivity().getIntent().getIntExtra("prospect", 0) + ";");
 
-                            Toast.makeText(getActivity(), "Cliente salvo com sucesso!", Toast.LENGTH_LONG).show();
-                            getActivity().finish();
+                                Toast.makeText(getActivity(), "Cliente salvo com sucesso!", Toast.LENGTH_LONG).show();
+                                getActivity().finish();
+
+                            } else {
+                                AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
+                                alert.setTitle("Atenção");
+                                switch (ClienteHelper.getCliente().getPessoa_f_j()) {
+                                    case "F":
+                                        alert.setMessage("Já existe outro cliente com esse CPF");
+                                        break;
+                                    case "J":
+                                        alert.setMessage("Já existe outro cliente com esse CNPJ");
+                                        break;
+                                    default:
+                                        alert.setMessage("Já existe outro cliente com esse CPF/CNPJ");
+                                        break;
+                                }
+                                alert.setNeutralButton("OK", new DialogInterface.OnClickListener() {
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int which) {
+                                        ClienteHelper.moveTela(0);
+                                    }
+                                });
+                                alert.show();
+                            }
                         }
                     } catch (Exception e) {
                         e.printStackTrace();
@@ -85,12 +112,18 @@ public class CadastroCliente8 extends Fragment {
 
         ClienteHelper.getCliente().setAtivo("S");
         ClienteHelper.getCliente().setId_empresa(Integer.parseInt(UsuarioHelper.getUsuario().getIdEmpresaMultiDevice()));
-        ClienteHelper.getCliente().setF_cliente("S");
+        ClienteHelper.getCliente().setF_cliente("N");
         ClienteHelper.getCliente().setF_fornecedor("N");
         ClienteHelper.getCliente().setF_funcionario("N");
         ClienteHelper.getCliente().setF_transportador("N");
         ClienteHelper.getCliente().setF_vendedor("N");
-        ClienteHelper.getCliente().setId_entidade("11");
+        ClienteHelper.getCliente().setId_entidade("10");
+    }
 
+    public boolean verificaCpfCnpj(String cpfCnpj) {
+        if (db.contagem("SELECT COUNT(*) FROM TBL_CADASTRO WHERE CPF_CNPJ = '" + cpfCnpj + "' AND ID_CADASTRO <> " + ClienteHelper.getCliente().getId_cadastro() + ";") > 0)
+            return false;
+        else
+            return true;
     }
 }
